requirements.txt
aiomysql==0.1.1
aiosqlite==0.17.0
anyio==3.6.2
bcrypt==4.0.1
click==8.1.3
colorama==0.4.6
fastapi==0.94.1
h11==0.14.0
httptools==0.5.0
idna==3.4
importlib-metadata==6.0.0
iso8601==1.1.0
passlib==1.7.4
pydantic==1.10.6
PyJWT==2.6.0
PyMySQL==1.0.2
pypika-tortoise==0.1.6
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
sniffio==1.3.0
starlette==0.26.1
tortoise-orm==0.19.3
typing_extensions==4.5.0
uvicorn==0.21.1
watchfiles==0.18.1
websockets==10.4
zipp==3.15.0
